moved reset-in-memory-repo-extension
