+++ domain/src/testFixtures/java/codes/fdk/blueprint/api/domain/stub/ResetInMemoryRepoExtension.java
@@ -1,6 +1,5 @@
-package codes.fdk.blueprint.api.infrastructure.rest.webflux;
+package codes.fdk.blueprint.api.domain.stub;
 
-import codes.fdk.blueprint.api.domain.stub.InMemoryCategoryRepository;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
